fto_analyzer_project/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
langchain==0.0.250
weaviate-client
pydantic==1.10.9
pytest==7.4.2
